+++ mbircone/src/interface.h
@@ -1,7 +1,7 @@
 #ifndef _CY_INTERFACE_H_
 #define _CY_INTERFACE_H_
 
-void AmatrixComputeToFile(double *angles, 
+void AmatrixComputeToFile(float *angles, 
 	struct SinoParams sinoParams, struct ImageParams imgParams, 
 	char *Amatrix_fname, char verbose);
